Lua 스크립트 연동

+++ Client/Client/Obj.h
@@ -29,7 +29,7 @@ protected:
 };
 
 
-inline int RequireExp(int HP) { return 50 * pow(2, HP); }
+inline int RequireExp(int HP) { return static_cast<int>(50 * pow(2, HP)); }
 inline int MaxHP(int Level) { return Level * 100; }
 
 class DynamicObj : public StaticObj
